+++ authority-portal-frontend/src/app/app.module.ts
@@ -26,18 +26,21 @@ import {AuthorityConnectorDetailPageModule} from './pages/authority-connector-de
 import {AuthorityConnectorListPageModule} from './pages/authority-connector-list-page/authority-connector-list-page.module';
 import {AuthorityOrganizationDetailPageModule} from './pages/authority-organization-detail-page/authority-organization-detail-page.module';
 import {AuthorityOrganizationListPageModule} from './pages/authority-organization-list-page/authority-organization-list-page.module';
-import {CentralComponentCreatePageModule} from './pages/central-component-create-page/central-component-create-page.module';
 import {CentralComponentListPageModule} from './pages/central-component-list-page/central-component-list-page.module';
 import {ProvideConnectorPageModule} from './pages/component-registration-pages/provide-connector-page/provide-connector-page.module';
+import {RegisterCentralComponentPageModule} from './pages/component-registration-pages/register-central-component-page/register-central-component-page.module';
+import {RegisterConnectorPageModule} from './pages/component-registration-pages/register-connector-page/register-connector-page.module';
+import {RequestConnectorPageModule} from './pages/component-registration-pages/request-connector-page/request-connector-page.module';
 import {ControlCenterPageModule} from './pages/control-center-page/control-center-page.module';
 import {DashboardPageModule} from './pages/dashboard-page/dashboard-page.module';
 import {ErrorPageModule} from './pages/empty-pages/error-page/error-page.module';
 import {LoadingPageModule} from './pages/empty-pages/loading-page/loading-page.module';
 import {PageNotFoundPageModule} from './pages/empty-pages/page-not-found-page/page-not-found-page.module';
 import {UnauthenticatedPageModule} from './pages/empty-pages/unauthenticated-page/unauthenticated-page.module';
+import {ParticipantCreateConnectorChoicePageModule} from './pages/participant-create-connector-choice-page/participant-create-connector-choice-page.module';
+import {ParticipantCreateProvidedConnectorChoicePageModule} from './pages/participant-create-provided-connector-choice-page/participant-create-provided-connector-choice-page.module';
 import {ParticipantOwnConnectorDetailPageModule} from './pages/participant-own-connector-detail-page/participant-own-connector-detail-page.module';
 import {ParticipantOwnConnectorListPageModule} from './pages/participant-own-connector-list-page/participant-own-connector-list-page.module';
-import {ParticipantRegisterOwnConnectorPageModule} from './pages/participant-register-own-connector-page/participant-register-own-connector-page.module';
 import {OrganizationCreatePageModule} from './pages/registration-pages/organization-create-page/organization-create-page.module';
 import {OrganizationOnboardPageModule} from './pages/registration-pages/organization-onboard-page/organization-onboard-page.module';
 import {OrganizationPendingPageModule} from './pages/registration-pages/organization-pending-page/organization-pending-page.module';
@@ -75,7 +78,6 @@ import {ParticipantInviteNewUserModule} from './popups/participant-invite-new-us
     AuthorityOrganizationDetailPageModule,
     AuthorityOrganizationListPageModule,
     AuthorityOrganizationUserDetailPageModule,
-    CentralComponentCreatePageModule,
     CentralComponentListPageModule,
     ControlCenterPageModule,
     DashboardPageModule,
@@ -92,10 +94,14 @@ import {ParticipantInviteNewUserModule} from './popups/participant-invite-new-us
     ParticipantOrganizationProfileModule,
     ParticipantOwnConnectorDetailPageModule,
     ParticipantOwnConnectorListPageModule,
-    ParticipantRegisterOwnConnectorPageModule,
+    ParticipantCreateConnectorChoicePageModule,
+    ParticipantCreateProvidedConnectorChoicePageModule,
     ParticipantUserDetailPageModule,
     PortalLayoutModule,
     ProvideConnectorPageModule,
+    RegisterConnectorPageModule,
+    RegisterCentralComponentPageModule,
+    RequestConnectorPageModule,
     SpConnectorDetailPageModule,
     SpConnectorListPageModule,
     ToastNotificationsModule,