+++ authority-portal-frontend/src/app/pages/component-registration-pages/provide-connector-page/state/provide-connector-page-state.ts
@@ -2,14 +2,12 @@ import {OrganizationOverviewEntryDto} from '@sovity.de/authority-portal-client';
 import {Fetched} from 'src/app/core/utils/fetched';
 
 export interface ProvideConnectorPageState {
-  connectorId: string;
   state: 'editing' | 'submitting' | 'success' | 'error';
   organizationList: Fetched<OrganizationOverviewEntryDto[]>;
   connectorConfig: string;
 }
 
-export const DEFAULT_PROVIDE_CONNECTOR_STATE: ProvideConnectorPageState = {
-  connectorId: '',
+export const DEFAULT_PROVIDE_CONNECTOR_PAGE_STATE: ProvideConnectorPageState = {
   state: 'editing',
   organizationList: Fetched.empty(),
   connectorConfig: '',