+++ authority-portal-frontend/src/app/pages/component-registration-pages/register-central-component-page/state/register-central-component-page-actions.ts
@@ -1,0 +1,18 @@
+import {CentralComponentCreateRequest} from '@sovity.de/authority-portal-client';
+
+const tag = 'RegisterCentralComponentPage';
+
+export class Reset {
+  static readonly type = `[${tag}] Reset`;
+}
+
+export class Submit {
+  static readonly type = `[${tag}] Register Connector`;
+
+  constructor(
+    public request: CentralComponentCreateRequest,
+    public enableForm: () => void,
+    public disableForm: () => void,
+    public success: () => void,
+  ) {}
+}