+++ authority-portal-frontend/src/app/pages/component-registration-pages/register-central-component-page/state/register-central-component-page-state.ts
@@ -1,0 +1,10 @@
+export interface RegisterCentralComponentPageState {
+  state: 'editing' | 'submitting' | 'success' | 'error';
+  createdCentralComponentId: string | null;
+}
+
+export const DEFAULT_REGISTER_CENTRAL_COMPONENT_PAGE_STATE: RegisterCentralComponentPageState =
+  {
+    state: 'editing',
+    createdCentralComponentId: null,
+  };