+++ authority-portal-frontend/src/app/pages/component-registration-pages/register-connector-page/state/register-connector-page-actions.ts
@@ -1,0 +1,17 @@
+import {CreateConnectorRequest} from '@sovity.de/authority-portal-client';
+
+const tag = 'RegisterConnectorPage';
+
+export class Reset {
+  static readonly type = `[${tag}] Reset`;
+}
+
+export class Submit {
+  static readonly type = `[${tag}] Register Connector`;
+  constructor(
+    public request: CreateConnectorRequest,
+    public enableForm: () => void,
+    public disableForm: () => void,
+    public success: () => void,
+  ) {}
+}