+++ authority-portal-frontend/src/app/pages/component-registration-pages/register-connector-page/state/register-connector-page-state.ts
@@ -1,0 +1,10 @@
+export interface RegisterConnectorPageState {
+  state: 'editing' | 'submitting' | 'success' | 'error';
+  connectorConfig: string;
+}
+
+export const DEFAULT_REGISTER_CONNECTOR_PAGE_STATE: RegisterConnectorPageState =
+  {
+    state: 'editing',
+    connectorConfig: '',
+  };