+++ authority-portal-frontend/src/app/pages/registration-pages/organization-onboard-page/state/organization-onboard-page-state.ts
@@ -2,12 +2,12 @@ import {OwnOrganizationDetailsDto} from '@sovity.de/authority-portal-client';
 import {Fetched} from 'src/app/core/utils/fetched';
 
 export interface OrganizationOnboardPageState {
-  onboardingFormState: 'editing' | 'submitting' | 'success' | 'error';
+  state: 'editing' | 'submitting' | 'success' | 'error';
   organization: Fetched<OwnOrganizationDetailsDto>;
 }
 
 export const DEFAULT_ORGANIZATION_ONBOARD_PAGE_PAGE_STATE: OrganizationOnboardPageState =
   {
-    onboardingFormState: 'editing',
+    state: 'editing',
     organization: Fetched.empty(),
   };